requirements.txt
# Dependencies for Git Autobot
GitPython>=3.1.40
PyGithub>=1.59.0
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
